backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
mysqlclient==2.2.4
django-cors-headers==4.3.1
pandas
joblib
scikit-learn
xgboost
